+++ src/Senparc.Weixin.Work/Senparc.Weixin.Work/AdvancedAPIs/External/ExternalJson/AddStrategyTagResult.cs
@@ -31,11 +31,11 @@ namespace Senparc.Weixin.Work.AdvancedAPIs.External.ExternalJson
         /// <summary>
         /// 标签组创建时间
         /// </summary>
-        public int create_time { get; set; }
+        public long create_time { get; set; }
         /// <summary>
         /// 标签组次序值。order值大的排序靠前。有效的值范围是[0, 2^32)
         /// </summary>
-        public int order { get; set; }
+        public long order { get; set; }
         /// <summary>
         /// 标签组内的标签列表
         /// </summary>
@@ -55,11 +55,11 @@ namespace Senparc.Weixin.Work.AdvancedAPIs.External.ExternalJson
         /// <summary>
         /// 标签创建时间
         /// </summary>
-        public int create_time { get; set; }
+        public long create_time { get; set; }
         /// <summary>
         /// 标签次序值。order值大的排序靠前。有效的值范围是[0, 2^32)
         /// </summary>
-        public int order { get; set; }
+        public long order { get; set; }
     }
 
 }
